pyproject.toml
[project]
name = "appointment-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.49.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "openai>=2.22.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
    "streamlit>=1.54.0",
]
